Extract API logic from odoo models. Create an AbstractModel for embedding this logic
