+++ setup/shopinvader_sale_cart/odoo/addons/shopinvader_sale_cart
@@ -1,0 +1,1 @@
+../../../../shopinvader_sale_cart
