+++ shopinvader_api_cart/models/__init__.py
@@ -1,2 +1,1 @@
 from . import sale_order
-from . import sale_order_line
