+++ shopinvader_sale_cart/models/__init__.py
@@ -1,0 +1,2 @@
+from . import sale_order
+from . import sale_order_line
