+++ shopinvader_sale_cart/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Marie Lejeune <marie.lejeune@acsone.eu>
