+++ shopinvader_sale_cart/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+This addon adds shopinvader logic on sale carts.
