+++ shopinvader_sale_cart/views/sale_order.xml
@@ -4,7 +4,6 @@
 <odoo>
 
     <record model="ir.ui.view" id="sale_order_form_view">
-        <field name="name">sale.order.form (in shopinvader_cart)</field>
         <field name="model">sale.order</field>
         <field name="inherit_id" ref="sale.view_order_form" />
         <field name="arch" type="xml">
